Extract `Token` and key traits to separate modules
